Add fine payment and fix bugs
