+++ docker-compose.yaml
@@ -10,11 +10,11 @@ services:
       - "8000:8000"
     command: >
       sh -c "python manage.py wait_for_db &&
+             python manage.py collectstatic --noinput &&
              python manage.py migrate &&
              gunicorn config.wsgi:application --bind 0.0.0.0:8000 --workers 3 --threads 4 --timeout 60 --graceful-timeout 30 --log-level info"
     volumes:
       - ./:/app
-      - my_media:/files/media
     depends_on:
       - db
 
@@ -113,6 +113,5 @@ services:
 
 volumes:
   my_db:
-  my_media:
   redis_data_drf:
   celery-schedule:
