+++ requirements.txt
@@ -1,18 +1,25 @@
+amqp==5.3.1
 asgiref==3.9.1
 attrs==25.3.0
+billiard==4.2.2
 black==25.9.0
 celery==5.5.3
 certifi==2025.8.3
 charset-normalizer==3.4.3
 click==8.3.0
+click-didyoumean==0.3.1
+click-plugins==1.1.1.2
+click-repl==0.3.0
 colorama==0.4.6
 coverage==7.10.6
+cron_descriptor==2.0.6
 Django==5.2.6
 django-celery-beat==2.8.1
 django-cors-headers==4.9.0
 django-debug-toolbar==6.0.0
 django-filter==25.1
 django-redis==6.0.0
+django-timezone-field==7.1
 djangorestframework==3.16.1
 djangorestframework_simplejwt==5.5.1
 drf-spectacular==0.28.0
@@ -21,11 +28,13 @@ flake8-quotes==3.4.0
 flake8-variables-names==0.0.6
 flower==2.0.1
 gunicorn==23.0.0
+humanize==4.13.0
 idna==3.10
 inflection==0.5.1
 isort==6.0.1
 jsonschema==4.25.1
 jsonschema-specifications==2025.9.1
+kombu==5.5.4
 mccabe==0.7.0
 mypy_extensions==1.1.0
 packaging==25.0
@@ -33,24 +42,34 @@ pathspec==0.12.1
 pep8-naming==0.15.1
 pillow==11.3.0
 platformdirs==4.4.0
+prometheus_client==0.23.1
+prompt_toolkit==3.0.52
 psycopg==3.2.10
 psycopg-binary==3.2.10
 pycodestyle==2.14.0
 pyflakes==3.4.0
 PyJWT==2.10.1
 pyTelegramBotAPI==4.29.1
+python-crontab==3.3.0
+python-dateutil==2.9.0.post0
 python-dotenv==1.1.1
 pytokens==0.1.10
+pytz==2025.2
 PyYAML==6.0.2
 redis==6.4.0
 referencing==0.36.2
 requests==2.32.5
 rpds-py==0.27.1
 setuptools==80.9.0
+six==1.17.0
 sqlparse==0.5.3
 stripe==12.5.1
 telebot==0.0.5
+tornado==6.5.2
 typing_extensions==4.15.0
 tzdata==2025.2
 uritemplate==4.2.0
 urllib3==2.5.0
+vine==5.1.0
+wcwidth==0.2.14
+whitenoise==6.11.0
